added in pretty_path method for the file system object
